refactor: 도메인 모델을 하위 패키지로 세분화
