+++ app/src/main/java/woowacourse/shopping/data/local/repository/RecentlyViewedProductRepositoryImpl.kt
@@ -4,7 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import woowacourse.shopping.data.local.dao.RecentlyViewedProductDao
 import woowacourse.shopping.data.local.entity.RecentlyViewedProductEntity
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.product.Product
 import woowacourse.shopping.domain.repository.RecentlyViewedProductRepository
 
 class RecentlyViewedProductRepositoryImpl(
