+++ app/src/main/java/woowacourse/shopping/data/remote/mock/dto/ResponseMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.remote.mock.dto
 
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.product.Product
 
 fun WebServerResponse.toObject() =
     Product(
