+++ app/src/main/java/woowacourse/shopping/data/remote/server/dto/cart/items/ContentResponse.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.remote.server.dto.cart.items
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import woowacourse.shopping.domain.model.PurchaseProduct
+import woowacourse.shopping.domain.model.order.PurchaseProduct
 
 @Serializable
 data class ContentResponse(
