+++ app/src/main/java/woowacourse/shopping/data/remote/server/dto/product/ProductResponse.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.remote.server.dto.product
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.product.Product
 
 @Serializable
 data class ProductResponse(
