--- app/src/main/java/woowacourse/shopping/domain/model/CartPage.kt
@@ -1,6 +1,0 @@
-package woowacourse.shopping.domain.model
-
-data class CartPage(
-    val items: PurchaseProducts,
-    val isLast: Boolean,
-)
