+++ app/src/main/java/woowacourse/shopping/domain/model/cart/CartPage.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.domain.model.cart
+
+import woowacourse.shopping.domain.model.order.PurchaseProducts
+
+data class CartPage(
+    val items: PurchaseProducts,
+    val isLast: Boolean,
+)
