+++ app/src/main/java/woowacourse/shopping/domain/model/product/Products.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.model
+package woowacourse.shopping.domain.model.product
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
