+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.domain.repository
 
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.product.Product
 
 interface ProductRepository {
     suspend fun getProducts(
