+++ app/src/main/java/woowacourse/shopping/domain/repository/RecentlyViewedProductRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.domain.repository
 
 import kotlinx.coroutines.flow.Flow
-import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.product.Product
 
 interface RecentlyViewedProductRepository {
     fun getAll(): Flow<List<Long>?>
