+++ app/src/test/java/woowacourse/shopping/data/remote/server/repository/CartRepositoryImplTest.kt
@@ -13,8 +13,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import retrofit2.Retrofit
 import woowacourse.shopping.data.remote.server.service.CartService
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.model.PurchaseProduct
+import woowacourse.shopping.domain.model.product.Product
+import woowacourse.shopping.domain.model.order.PurchaseProduct
 import woowacourse.shopping.domain.repository.CartRepository
 
 class CartRepositoryImplTest {
