+++ app/src/test/java/woowacourse/shopping/domain/model/product/ProductTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.model
+package woowacourse.shopping.domain.model.product
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
