+++ app/src/test/java/woowacourse/shopping/ui/catalog/ShoppingViewModelTest.kt
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
-import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.model.PurchaseProduct
+import woowacourse.shopping.domain.model.product.Product
+import woowacourse.shopping.domain.model.order.PurchaseProduct
 import woowacourse.shopping.testing.fakes.FakeCartRepository
 import woowacourse.shopping.testing.fakes.FakeProductRepository
 import woowacourse.shopping.testing.fakes.FakeRecentlyViewedProductRepository
